Remove legacy Victor-based driveline
Upgraded the Practice Bot chassis to use Talons.  This allows us to rip
out the legacy Victor driveline code and use the identical driveline
setup as the real robot (simplifying code).
This will also allow us to develop most of the "drive-related" code on
the practice bot while the real robot is under development.
Also added a saved pipeline file for the LimeLight based on some initial
tuning in the 'pits.