Position PID working.  Add basic elevator
Changed PID mode for driveline from motion magic to simpler position (As
a part of debugging).  Code refactoring and cleanup.  Safety causes
followers to halt and stutter, so turned off driveline safety for now.
Added hack because Talon is driving 2x farther than it should, even
though the print functions in the talon agree that it is going 2x
farther.  When we set the distance to be half of what we "should" want,
it goes the right distance.  To be debugged later.