Convert DriveTrain's SpeedControllerGroup to use just the Master talon
Discovered the slave stuttering was due to the motor safety firing.
Motor safety was firing because we listed both slave and follower inside
a SpeedControllerGroup (left over from when we had Victor888's on the
drivetrain and needed to control both motors explicitly).  However, in
follower mode we don't update the follower speeds (it's handled in
hardware) so the motor safety was firing.  Now, we just drive the
masters and let the slaves follow.

Note:  we don't explicitly set the followers to follower mode on init or
in the joystick control mode.... yet they are still following!  This
needs more investigation.